Patch updated and typo fixes for string

+++ android/src/oldarch/RnAndroidSmsRetrieverSpec.kt
@@ -7,8 +7,7 @@ import com.facebook.react.bridge.Promise
 abstract class RnAndroidSmsRetrieverSpec internal constructor(context: ReactApplicationContext) :
   ReactContextBaseJavaModule(context) {
 
-  abstract fun getOtp(otpLength: Int, phoneNumber: String?, promise: Promise)
+  abstract fun getOtp(otpLength: Double, phoneNumber: String?, promise: Promise)
 
-  abstract fun getSms(phoneNumber: String?,promise: Promise)
-
+  abstract fun getSms(phoneNumber: String?, promise: Promise)
 }
